Recompute subtree count on node addition

+++ ulyp-ui/src/main/java/com/ulyp/ui/CallTreeNodeContent.java
@@ -7,7 +7,7 @@ import javafx.scene.layout.StackPane;
 // TODO should be a better name
 public class CallTreeNodeContent extends StackPane {
 
-    public CallTreeNodeContent(CallRecord node, RenderSettings renderSettings, int totalNodeCountInTree) {
+    public CallTreeNodeContent(CallRecord node, RenderSettings renderSettings, long totalNodeCountInTree) {
 
         setAlignment(Pos.CENTER_LEFT);
         getChildren().addAll(new CallRecordTreeNodeRelativeWeight(node, totalNodeCountInTree), new RenderedCallRecord(node, renderSettings));
